terminado as exclusões

+++ app/Http/Controllers/ProdutoController.php
@@ -5,13 +5,16 @@ namespace App\Http\Controllers;
 use App\Models\Categoria;
 use App\Models\Produto;
 use App\Models\Estoque;
+use App\Models\Compra;
 use Illuminate\Http\Request;
 
-class ProdutoController extends Controller {
+class ProdutoController extends Controller
+{
     /**
      * Display a listing of the resource.
      */
-    public function index() {
+    public function index()
+    {
         $produtos = Produto::all();
         return view('adm.produto.index', compact('produtos'));
     }
@@ -19,7 +22,8 @@ class ProdutoController extends Controller {
     /**
      * Show the form for creating a new resource.
      */
-    public function create() {
+    public function create()
+    {
         $categorias = Categoria::all();
         return view('adm.produto.create', compact('categorias'));
     }
@@ -27,22 +31,25 @@ class ProdutoController extends Controller {
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request) {
+    public function store(Request $request)
+    {
         Produto::create($request->all());
-        return redirect()->back()->with('message','Cadastrado com sucesso!');
+        return redirect()->back()->with('message', 'Cadastrado com sucesso!');
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Produto $produto) {
+    public function show(Produto $produto)
+    {
         return view('adm.produto.show', compact('produto'));
     }
 
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Produto $produto) {
+    public function edit(Produto $produto)
+    {
         return view('adm.produto.edit');
     }
 
@@ -57,10 +64,16 @@ class ProdutoController extends Controller {
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Produto $produto)
+    public function destroy($id_produto)
     {
-        $produto = Produto::findOrFail($produto->id_produto);
-        $produto->delete();
-        return redirect()->route('adm.produto.index')->with('message','Deletado com sucesso!');
+        Compra::where('id_produto', $id_produto)->delete();
+
+        $produto = Produto::find($id_produto);
+        if ($produto) {
+            $produto->delete();
+            return redirect()->route('adm.produto.index')->with('message', 'Deletado com sucesso!');
+        } else {
+            return redirect()->route('adm.produto.index')->with('error', 'Produto não encontrado!');
+        }
     }
 }
